# lib/api-zod/src/generated/types/jobItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobItem {
  id: string;
  title: string;
  organization: string;
  jobType: string;
  state: string;
  district: string;
  lastDate: string;
  vacancies: number;
  qualification: string;
  salaryMin: number;
  salaryMax: number;
  applyUrl: string;
  description: string;
  eligibility: string;
  isNew: boolean;
  postedDate: string;
  category: string;
}