# lib/api-zod/src/generated/types/jobsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobItem } from "./jobItem";

export interface JobsResponse {
  jobs: JobItem[];
  total: number;
  page: number;
  totalPages: number;
  lastFetched: string;
}